+++ src/compiler/message.rs
@@ -221,10 +221,11 @@ impl Problem {
       Self::AccessPathStartsWithIndex =>                        rcode!(0021),
       Self::AccessPathStartsWithSlice =>                        rcode!(0022),
       Self::InvalidSliceBound(_) =>                             rcode!(0023),
-      Self::NothingToPipe =>                                 rcode!(0024),
+      Self::NothingToPipe =>                                    rcode!(0024),
       Self::DynamicDepth =>                                     rcode!(0025),
       Self::DepthAssignment =>                                  rcode!(0026),
       Self::InvalidDepthUsage =>                                rcode!(0027),
+      Self::FallibleOptionalArgAccess(_) =>                     rcode!(0028),
       
       // Static analysis errors (0100 - 0199)
       Self::ConstantReassignment(_) =>                          rcode!(0100),
@@ -244,7 +245,6 @@ impl Problem {
       Self::UnusedFunction(_) =>                                rcode!(1002),
       Self::EmptyFunctionBody(_) =>                             rcode!(1003),
       Self::NestedFunctionDefMarkedConstant =>                  rcode!(1004),
-      Self::FallibleOptionalArgAccess(_) =>                     rcode!(1005),
       
       // File access errors (0100 - 0109)
       Self::FileNotFound(_) =>                                  rcode!(2100),
@@ -296,7 +296,7 @@ impl Problem {
       Self::DynamicDepth => rmsg!("depth operator cannot be used on dynamic variable names"),
       Self::DepthAssignment => rmsg!("variable depth cannot be assigned to"),
       Self::InvalidDepthUsage => rmsg!("depth operator is not valid in this context"),
-      Self::FallibleOptionalArgAccess(argname) => rmsg!("access to optional argument '{}' can fail; consider adding a fallback to the accessor or specifying a default argument", argname),
+      Self::FallibleOptionalArgAccess(argname) => rmsg!("access to optional argument '{}' can fail; add a fallback to the accessor or specify a default argument", argname),
     }
   }